getDevices refactoring
